move cert generator code into certificate_test.go and introduce Identity.IsCertificateBased (#39)
* introduce Identity.IsCertificateBased
* move cert generator code into certificate_test.go
