+++ auth/proofOfPossession_test.go
@@ -4,16 +4,8 @@
 package auth
 
 import (
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/base64"
-	"encoding/pem"
-	"github.com/lestrrat-go/jwx/jwt"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"math/big"
 	"testing"
 )
 
@@ -112,46 +104,4 @@ func TestProofOfPossession_validateX5tThumbprint(t *testing.T) {
 			}
 		})
 	}
-}
-
-func generateToken(t *testing.T, claimCnfMemberX5tValue string) Token {
-	token := jwt.New()
-	cnfClaim := map[string]interface{}{
-		claimCnfMemberX5t: claimCnfMemberX5tValue,
-	}
-	err := token.Set(claimCnf, cnfClaim)
-	require.NoError(t, err, "Failed to create token: %v", err)
-
-	return stdToken{jwtToken: token}
-}
-
-func convertToPEM(t *testing.T, derCert string) string {
-	x509Cert, err := newCertificate(derCert)
-	require.NoError(t, err, "failed to create certificate: %v", err)
-
-	bytes := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: x509Cert.x509Cert.Raw})
-	return base64.StdEncoding.EncodeToString(bytes)
-}
-
-func generateDERCert() string {
-	key, _ := rsa.GenerateKey(rand.Reader, 512) //nolint:gosec
-
-	issuerName := pkix.Name{
-		Organization: []string{"my-issuer-org"},
-	}
-	template := x509.Certificate{
-		SerialNumber: big.NewInt(125),
-		Subject: pkix.Name{
-			Organization: []string{"my-subject-org"},
-		},
-		Issuer: issuerName,
-	}
-	issTemplate := x509.Certificate{
-		SerialNumber: big.NewInt(12345),
-		Subject:      issuerName,
-		Issuer:       issuerName,
-	}
-	derBytes, _ := x509.CreateCertificate(rand.Reader, &template, &issTemplate, &key.PublicKey, key)
-
-	return base64.StdEncoding.EncodeToString(derBytes)
 }
